Add custom hook for handling oauth authentication
