+++ app/src/views/SignUp.tsx
@@ -9,63 +9,38 @@ import {
 } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 
-import {useNavigate, useLocation, useParams} from 'react-router-dom';
-import {setCredentials} from '../app/authSlice';
-import {useAppDispatch} from '../app/hooks/store';
+import {useNavigate, useParams} from 'react-router-dom';
 import {SignUpForm} from '../components/authentication/SignUpForm';
 import {
   SignUpHostRequest,
   SignUpCoordinatorRequest,
   useSignUpHostMutation,
   useSignUpCoordinatorMutation,
-  useGetTokenMutation,
 } from '../services/auth';
 import {isErrorWithMessage, isFetchBaseQueryError} from '../app/helpers';
 import {FormContainer} from '../components/authentication';
+import {useAuthenticateWithOAuth} from '../components/authentication/hooks/useAuthenticateWithOAuth';
 
 export const SignUp = () => {
   const [errorMessage, setErrorMessage] = React.useState('');
   const {type} = useParams();
   const navigate = useNavigate();
-  const location = useLocation();
-  const dispatch = useAppDispatch();
 
   const [signUpHost, {isLoading: signUpHostIsLoading}] =
     useSignUpHostMutation();
   const [signUpCoordinator, {isLoading: signUpCoordinatorIsLoading}] =
     useSignUpCoordinatorMutation();
-  const [getToken, {isLoading: getTokenIsLoading}] = useGetTokenMutation();
   // get type from params
   // const locationState = location.state as LocationState;
 
   // Save location from which user was redirected to login page
   // const from = locationState?.from?.pathname || '/';
 
-  React.useEffect(() => {
-    if (location.search.includes('code')) {
-      const code = location.search.split('?code=')[1];
-      getToken({
-        code,
-        callbackUri: `/signup/${type}`,
-      })
-        .unwrap()
-        .then(response => {
-          const {token, user} = response;
-          dispatch(setCredentials({user, token}));
-          navigate('/coordinator');
-        })
-        .catch(err => {
-          if (isFetchBaseQueryError(err)) {
-            // you can access all properties of `FetchBaseQueryError` here
-            const errMsg = err.data.message;
-            setErrorMessage(errMsg);
-          } else if (isErrorWithMessage(err)) {
-            // you can access a string 'message' property here
-            setErrorMessage(err.message);
-          }
-        });
-    }
-  }, [location]);
+  const callbackUri = `/signup/${type}`;
+  const {getTokenIsLoading} = useAuthenticateWithOAuth({
+    setErrorMessage,
+    callbackUri,
+  });
 
   const handleSignUp = async ({
     email,
